feat: add Has() function to check the file exist or not, attach unit test

+++ storage/storage_test.go
@@ -8,29 +8,21 @@ import (
 )
 
 func TestPathTransformFunc(t *testing.T) {
-	key := "SydneyPic"
+	key := "SydneyHoliday"
 	pathKey := CASPathTransformFunc(key)
 	fmt.Println(pathKey.FileName)
 	fmt.Println(pathKey.fullPath())
 }
 
 func TestStorage(t *testing.T) {
-	opts := StorageOpt{
-		PathTransformFunc: CASPathTransformFunc,
-	}
-	store := NewStore(opts)
-
-	key := "SydneyPic"
-	data := []byte("some jpg file byes")
-	if err := store.writeStream(key, bytes.NewReader(data)); err != nil {
-		t.Error(err)
-	}
-
+	// init
+	key := "SydneyHoliday"
+	store, data := storeFile(key, t)
+	// read file
 	r, err := store.Read(key)
 	if err != nil {
 		t.Error(err)
 	}
-
 	// 读取并使用String比对
 	b, _ := io.ReadAll(r)
 	fmt.Println(string(b))
@@ -40,15 +32,34 @@ func TestStorage(t *testing.T) {
 }
 
 func TestStorage_Delete(t *testing.T) {
-	opts := StorageOpt{
-		PathTransformFunc: CASPathTransformFunc,
+	key := "SydneyHoliday"
+	store, _ := storeFile(key, t)
+	// check exist
+	if exist := store.Has(key); !exist {
+		t.Errorf("Does not exist file %s\n", key)
+		return
 	}
-	store := NewStore(opts)
-
 	// delete
-	key := "SydneyPic"
 	err := store.Delete(key)
 	if err != nil {
 		t.Error(err)
 	}
+	// check exist
+	if exist := store.Has(key); exist {
+		t.Errorf("Does not delete the file %s\n", key)
+		return
+	}
+}
+
+// storeFile 测试使用, 保存文件
+func storeFile(key string, t *testing.T) (*Storage, []byte) {
+	opts := StorageOpt{
+		PathTransformFunc: CASPathTransformFunc,
+	}
+	store := NewStore(opts)
+	data := []byte("some jpg file byes")
+	if err := store.writeStream(key, bytes.NewReader(data)); err != nil {
+		t.Error(err)
+	}
+	return store, data
 }
